server/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
web3==6.11.3
pydantic==2.5.0
python-multipart==0.0.6
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
